requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Session==0.3.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
#pkg-resources==0.0.0
PyMuPDF==1.16.8
PyPDF2==1.26.0
reportlab==3.5.32
requests==2.22.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
